[Refactor] STAMPIT-161 Mapper 분리
- HomeViewModel
