+++ StampIt-Project/StampIt-Project/Presentation/Common/Mapper/Protocol/MemberMapping.swift
@@ -1,0 +1,12 @@
+//
+//  MemberMapping.swift
+//  StampIt-Project
+//
+//  Created by daeun on 6/24/25.
+//
+
+import Foundation
+
+protocol MemberMapping {
+    func map(members: [Member], userID: String) -> [HomeItem]
+}
